TV: Move HomeViewModel to its own file
